refactor tests locations
